OTE migration for kdo
formatted
